% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{cv.multiFSSEMiPALM}
\alias{cv.multiFSSEMiPALM}
\title{cv.multiFSSEMiPALM}
\usage{
cv.multiFSSEMiPALM(Xs, Ys, Bs, Fs, Sk, sigma2, nlambda = 20, nrho = 20,
  nfold = 5, p, q, wt = TRUE, plot = FALSE)
}
\arguments{
\item{Xs}{eQTL matrices}

\item{Ys}{Gene expression matrices}

\item{Bs}{initialized GRN-matrices}

\item{Fs}{initialized eQTL effect matrices}

\item{Sk}{eQTL index of genes}

\item{sigma2}{initialized noise variance}

\item{nlambda}{number of hyper-parameter of lasso term in CV}

\item{nrho}{number of hyper-parameter of fused-lasso term in CV}

\item{nfold}{CVfold number. Default 5/10}

\item{p}{number of genes}

\item{q}{number of eQTLs}

\item{wt}{use adaptive lasso or not. Default TRUE.}

\item{plot}{plot contour of cvmean or not. Default FALSE.}
}
\value{
list of cross-validation result
}
\description{
cv.multiFSSEMiPALM
}
