% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{cv.multiRegression}
\alias{cv.multiRegression}
\title{cv.multiRegression}
\usage{
cv.multiRegression(Xs, Ys, Sk, ngamma = 20, nfold = 5, n, p, k)
}
\arguments{
\item{Xs}{eQTL matrices}

\item{Ys}{Gene expression matrices}

\item{Sk}{eQTL index of genes}

\item{ngamma}{number of hyper-parameter in CV}

\item{nfold}{CVfold number. Default 5/10}

\item{n}{number of observations}

\item{p}{number of genes}

\item{k}{number of eQTLs}
}
\value{
gamma_min optimal gamma to minimize cross-validation error
}
\description{
cv.multiRegression
}
