% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cwiseGradient4FSSEM}
\alias{cwiseGradient4FSSEM}
\title{cwiseGradient4FSSEM}
\usage{
cwiseGradient4FSSEM(n, c, Y, R, Y2norm, sigma2)
}
\arguments{
\item{n}{number of observations}

\item{c}{cofactor vector}

\item{Y}{Matrix of gene expression}

\item{R}{Residual matrix}

\item{Y2norm}{Column of YtY}

\item{sigma2}{noise variance}
}
\value{
function whose argument is column vector bi
}
\description{
function generator function
}
