% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FDR}
\alias{FDR}
\title{FDR}
\usage{
FDR(X, B, PREC = 0)
}
\arguments{
\item{X}{list of predicted network matrices}

\item{B}{list of true network matrices}

\item{PREC}{precision threshold for FDR test. Default 0.}
}
\description{
False discovery rate for network prediction
}
