% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flinvB}
\alias{flinvB}
\title{flinvB}
\usage{
flinvB(Bs)
}
\arguments{
\item{Bs}{list of network matrices}
}
\value{
inversed difference matrices
}
\description{
inversed difference of two B matrices. For adaptive fused lasso penalty
}
