% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{floneB}
\alias{floneB}
\title{floneB}
\usage{
floneB(Bs)
}
\arguments{
\item{Bs}{list of network matrices}
}
\value{
matrix whose entries are all 1
}
\description{
if you do not want adaptive fused lasso penalty, floneB replace flinvB
}
