% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{implFSSEM}
\alias{implFSSEM}
\title{implFSSEM}
\usage{
implFSSEM(data = NULL, method = c("CV", "BIC"))
}
\arguments{
\item{data}{Data archive of experiment measurements, includeing
eQTL matrices, Gene expression matrices of different conditions,
marker of eQTLs and data generation SEM model}

\item{method}{Use cross-validation (CV) or bayesian-information-criterion(BIC)}
}
\value{
List of TPR and FDR
}
\description{
implementor function of FSSEM solver
}
