% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssem.R
\name{initLambdaiPALM2}
\alias{initLambdaiPALM2}
\title{initLambdaiPALM2}
\usage{
initLambdaiPALM2(Xs, Ys, Bs, Fs, Sk, sigma2, Wl, Wf, p, k)
}
\arguments{
\item{Xs}{eQTL matrices}

\item{Ys}{Gene expression matrices}

\item{Bs}{initialized GRN-matrices}

\item{Fs}{initialized eQTL effect matrices}

\item{Sk}{eQTL index of genes}

\item{sigma2}{initialized noise variance}

\item{Wl}{weight matrices for adaptive lasso terms}

\item{Wf}{weight matrix for adaptive fused lasso term}

\item{p}{number of genes}

\item{k}{number of eQTL}
}
\value{
lambda_max
}
\description{
initLambdaiPALM2
}
