% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssem.R
\name{initRhoiPALM2}
\alias{initRhoiPALM2}
\title{initRhoiPALM2}
\usage{
initRhoiPALM2(Xs, Ys, Bs, Fs, Sk, sigma2, Wl, Wf, lambda, n, p)
}
\arguments{
\item{Xs}{eQTL matrices}

\item{Ys}{Gene expression matrices}

\item{Bs}{initialized GRN-matrices}

\item{Fs}{initialized eQTL effect matrices}

\item{Sk}{eQTL index of genes}

\item{sigma2}{initialized noise variance}

\item{Wl}{weight matrices for adaptive lasso terms}

\item{Wf}{weight matrix for adaptive fused lasso term}

\item{lambda}{lambda w.r.t. rho_max}

\item{n}{number of observations}

\item{p}{number of genes}
}
\value{
rho_max
}
\description{
initRhoiPALM2
}
