% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inverseB}
\alias{inverseB}
\title{inverseB}
\usage{
inverseB(Bs)
}
\arguments{
\item{Bs}{list of network matrices}
}
\value{
list of inversed B matrices
}
\description{
inverse matrices of B network for adaptive FSSEM
}
