% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invoneB}
\alias{invoneB}
\title{invoneB}
\usage{
invoneB(Bs)
}
\arguments{
\item{Bs}{list of network matrices}
}
\value{
list of invone B matrices
}
\description{
if you do not want to get inversed B matrces, invoneB gives you a matrix with constant 1 instead in FSSEM
}
