% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLikNFSSEM}
\alias{logLikNFSSEM}
\title{logLikNFSSEM}
\usage{
logLikNFSSEM(Bs, Wl, Wf, lambda, rho, sigma2, Dets, n, p)
}
\arguments{
\item{Bs}{Network matrices}

\item{Wl}{Weights for lasso term}

\item{Wf}{Weights for group perturb lasso term}

\item{lambda}{Hyperparameter of lasso term}

\item{rho}{Hyperparameter of group fused lasso term}

\item{sigma2}{noise variance}

\item{Dets}{determinants of I-B matrices}

\item{n}{number of observations}

\item{p}{number of genes}
}
\value{
objective value of NFSSEM with specified hyper-paramters
}
\description{
logLikNFSSEM
}
