% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{multiRegression}
\alias{multiRegression}
\title{multiRegression}
\usage{
multiRegression(Xs, Ys, Sk, gamma, n, p, k, trans = FALSE)
}
\arguments{
\item{Xs}{eQTL matrices. eQTL matrix can be matrix/list of multiple conditions}

\item{Ys}{Gene expression matrices}

\item{Sk}{eQTL index of genes}

\item{gamma}{Hyperparameter for ridge regression}

\item{n}{number of observations}

\item{p}{number of genes}

\item{k}{number of eQTLs}

\item{trans}{if rows for sample, trans = TRUE, otherwise, trans = FALSE. Default FALSE}
}
\value{
fit List of SEM model
\describe{
\item{Bs}{ coefficient matrices of gene regulatory networks}
\item{fs}{ eQTL's coefficients w.r.t each gene}
\item{Fs}{ coefficient matrices of eQTL-gene effect}
\item{mu}{ Bias vector}
\item{sigma2}{ estimate of covariance in SEM}
}
}
\description{
Ridge regression on multiple conditions
}
