% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obj.multiRegression}
\alias{obj.multiRegression}
\title{obj.multiRegression}
\usage{
obj.multiRegression(Xs, Ys, fit, trans = F)
}
\arguments{
\item{Xs}{eQTL matrices}

\item{Ys}{gene expression matrices}

\item{fit}{regression fit result object}

\item{trans}{if rows for sample, trans = TRUE, otherwise, trans = FALSE. Default FALSE}
}
\value{
error squared norm of ||(I-B)Y - FX||_2^2
}
\description{
obj.multiRegression
}
