% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssem.R
\name{opt.multiFSSEMiPALM2}
\alias{opt.multiFSSEMiPALM2}
\title{opt.multiFSSEMiPALM2}
\usage{
opt.multiFSSEMiPALM2(Xs, Ys, Bs, Fs, Sk, sigma2, nlambda = 20,
  nrho = 20, p, q, wt = TRUE)
}
\arguments{
\item{Xs}{eQTL matrices}

\item{Ys}{Gene expression matrices}

\item{Bs}{initialized GRN-matrices}

\item{Fs}{initialized eQTL effect matrices}

\item{Sk}{eQTL index of genes}

\item{sigma2}{initialized noise variance}

\item{nlambda}{number of hyper-parameter of lasso term in CV}

\item{nrho}{number of hyper-parameter of fused-lasso term in CV}

\item{p}{number of genes}

\item{q}{number of eQTLs}

\item{wt}{use adaptive lasso or not. Default TRUE.}
}
\value{
list of model selection result
}
\description{
optimize multiFSSEMiPALM's parameters by minimize BIC,
when feature size is large (> 300), BIC methods will be much faster than Cross-validation
}
