% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfssem.R
\name{opt.multiNFSSEMiPALM2}
\alias{opt.multiNFSSEMiPALM2}
\title{opt.multiNFSSEMiPALM2}
\usage{
opt.multiNFSSEMiPALM2(
  Xs,
  Ys,
  Bs,
  Fs,
  Sk,
  sigma2,
  nlambda = 20,
  nrho = 20,
  p,
  q,
  wt = TRUE
)
}
\arguments{
\item{Xs}{eQTL matrices}

\item{Ys}{Gene expression matrices}

\item{Bs}{initialized GRN-matrices}

\item{Fs}{initialized eQTL effect matrices}

\item{Sk}{eQTL index of genes}

\item{sigma2}{initialized noise variance}

\item{nlambda}{number of hyper-parameter of lasso term in CV}

\item{nrho}{number of hyper-parameter of fused-lasso term in CV}

\item{p}{number of genes}

\item{q}{number of eQTLs}

\item{wt}{use adaptive lasso or not. Default TRUE.}
}
\value{
list of model selection result
}
\description{
optimize multiNFSSEMiPALM's parameters by minimize BIC,
when feature size is large (> 300), BIC methods will be much faster than Cross-validation
}
