% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pninvB}
\alias{pninvB}
\title{pninvB}
\usage{
pninvB(Bs)
}
\arguments{
\item{Bs}{list of network matrices}
}
\value{
inversed l2 norm of B2 - B1
}
\description{
inversed column l2 norm for perturbed group lasso penalty
}
