% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pnoneB}
\alias{pnoneB}
\title{pnoneB}
\usage{
pnoneB(Bs)
}
\arguments{
\item{Bs}{list of network matrices}
}
\value{
inversed l2 norm of B2 - B1 with all entries is 1
}
\description{
if you do not want adaptive group lasso penalty, pnoneB replace pninvB
}
