% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proc.centerFSSEM}
\alias{proc.centerFSSEM}
\title{proc.centerFSSEM}
\usage{
proc.centerFSSEM(Xs, Ys)
}
\arguments{
\item{Xs}{eQTL matrices}

\item{Ys}{list of gene expression matrices}
}
\value{
centered Xs and Ys and mean vectors
}
\description{
proc.centerFSSEM
}
