% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proc.centerFSSEM2}
\alias{proc.centerFSSEM2}
\title{proc.centerFSSEM2}
\usage{
proc.centerFSSEM2(Xs, Ys)
}
\arguments{
\item{Xs}{list of eQTL matrices}

\item{Ys}{list of gene expression matrices}
}
\value{
centered Xs and Ys and mean vectors
}
\description{
proc.centerFSSEM2
}
