% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{randomFSSEMdata}
\alias{randomFSSEMdata}
\title{randomFSSEMdata}
\usage{
randomFSSEMdata(n, p, k, sparse = 0.1, df = 0.2, sigma2 = 0.01,
  u = 5, type = c("DG", "ER"), dag = TRUE, coef = c(0.2, 0.4),
  nhub = 2)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of genes}

\item{k}{number of eQTLs}

\item{sparse}{ratio of edges / gene_number}

\item{df}{ratio of differential edges among two network}

\item{sigma2}{noise variance of error}

\item{u}{variance of bias in SEM model.}

\item{type}{type of generated network, can be selected as DG, ER, Scale-free network}

\item{dag}{network is directed-acyclic or not. Default TRUE}

\item{coef}{Range of absolute value of coefficients in simulated network matrices. Default (0.2, 0.4), or (0.5, 1)}

\item{nhub}{If you select to generate ER network, nhub is the number of pre-defined hub node number. Default 2}
}
\value{
list of generated data
\describe{
\item{Data}{ List of observed, Xs, Ys, Sk }
\item{Vars}{ List of model, Bs, Fs, mu, n, p, k }
}
}
\description{
randomFSSEMdata
}
