% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TPR}
\alias{TPR}
\title{TPR}
\usage{
TPR(X, B, PREC = 0)
}
\arguments{
\item{X}{list of predicted network matrices}

\item{B}{list of true network matrices}

\item{PREC}{precision threshold for FDR test. Default 0.}
}
\description{
Power of detection for network prediction
}
