% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{transx}
\alias{transx}
\title{transx}
\usage{
transx(data)
}
\arguments{
\item{data}{Collecting data structure generated by randomFSSEMdata function}
}
\value{
transformed list of eQTL matrices
}
\description{
transx
}
